+++ pyaldata/__init__.py
@@ -10,3 +10,4 @@ from .cmp import *
 from .smoothing import *
 from .signal_transformations import *
 from .firing_rates import *
+from .df_utils import *
